heat map: refactor into different files & add tests.
